refactor: merge error and config libraries into base
- Combined error and config libraries into a unified base library.
- Simplifies dependency management and improves maintainability.
