+++ base/Cargo.toml
@@ -1,11 +1,16 @@
 [package]
-name = "error"
+name = "base"
 authors.workspace = true
 version = "0.1.0"
 edition.workspace = true
 license.workspace = true
 
 [dependencies]
+serde.workspace = true
+toml.workspace = true
+anyhow.workspace = true
+tracing.workspace = true
+tracing-subscriber.workspace = true
 thiserror.workspace = true
 regex.workspace = true
 reqwest.workspace = true
