--- config/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "config"
-authors.workspace = true
-version = "0.1.0"
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-serde.workspace = true
-toml.workspace = true
-anyhow.workspace = true
-tracing.workspace = true
-tracing-subscriber.workspace = true
