+++ telegram/Cargo.toml
@@ -13,4 +13,4 @@ serde_json.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 async-trait.workspace = true
-error.workspace = true
+base.workspace = true
