+++ wakatime/Cargo.toml
@@ -21,5 +21,4 @@ tracing.workspace = true
 tracing-subscriber.workspace = true
 
 # urdekcah
-config.workspace = true
-error.workspace = true
+base.workspace = true
