+++ weather/Cargo.toml
@@ -17,4 +17,4 @@ tracing-subscriber.workspace = true
 url.workspace = true
 
 # urdekcah
-error.workspace = true
+base.workspace = true
