+++ weather/src/config.rs
@@ -3,7 +3,7 @@
 // Этот исходный код распространяется под лицензией AGPL-3.0,
 // текст которой находится в файле LICENSE в корневом каталоге данного проекта.
 use anyhow::Result;
-use error::Error;
+use base::Error;
 use std::path::PathBuf;
 
 #[derive(Debug, Clone)]
